@@ -8,11 +8,17 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@ant-design/icons-vue": "^6.1.0",
+    "ant-design-vue": "^2.2.8",
     "axios": "^0.20.0",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11",
+    "qrcode.vue": "3.3.3",
+    "vue": "3.2.0",
+    "vue-cli-plugin-browser-extension": "^0.25.2",
+    "vue-loader": "^17.0.0",
     "vue-router": "^3.2.0",
-    "vuex": "^3.4.0"
+    "vuex": "4.0.2",
+    "webpack": "4.46.0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.15",
@@ -20,6 +26,7 @@
     "@vue/cli-plugin-router": "~4.5.15",
     "@vue/cli-plugin-vuex": "~4.5.15",
     "@vue/cli-service": "~4.5.15",
+    "@vue/compiler-sfc": "^3.2.33",
     "@vue/eslint-config-standard": "^5.1.2",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
@@ -29,8 +36,7 @@
     "eslint-plugin-standard": "^4.0.0",
     "eslint-plugin-vue": "^6.2.2",
     "sass": "^1.26.5",
-    "sass-loader": "^8.0.2",
-    "vue-cli-plugin-browser-extension": "latest",
+    "sass-loader": "10.2.1",
     "vue-template-compiler": "^2.6.11"
   }
 }
