@@ -1,8 +1,12 @@
-import Vue from 'vue'
-import App from './App.vue'
+import { createApp } from 'vue'
+// import Antd from 'ant-design-vue';
+import Antd from 'ant-design-vue'
+import App from './App'
+import 'ant-design-vue/dist/antd.css'
+import './popup.css'
 
-/* eslint-disable no-new */
-new Vue({
-  el: '#app',
-  render: h => h(App)
-})
+const app = createApp(App)
+app.config.productionTip = true
+
+app.use(Antd).mount('#app')
+app.config.globalProperties.$message = message
