@@ -1,12 +1,6 @@
-import Vue from 'vue'
 import App from './App.vue'
-import router from './router'
-import store from './store'
+import { createApp } from 'vue'
 
-Vue.config.productionTip = false
+const app = createApp(App)
 
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount('#app')
+app.use().mount('#app')
